Removing redundant usings.
Canging points cost to decimal.

+++ src/Entelect.Encentivize.Sdk/Rewards/Reward.cs
@@ -5,7 +5,7 @@ namespace Entelect.Encentivize.Sdk.Rewards
         public long RewardId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int PointsCost { get; set; }
+        public decimal PointsCost { get; set; }
         public string TermsAndConditions { get; set; }
         public string PrimaryImageUrl { get; set; }
     }
